R Code
QUESTION

Data




For this analysis you will use the loans data available from here:




https://spark-public.s3.amazonaws.com/dataanalysis/loansData.csv
https://spark-public.s3.amazonaws.com/dataanalysis/loansData.rda




There is a code book for the variables in the data set available here:




https://spark-public.s3.amazonaws.com/dataanalysis/loansCodebook.pdf




Prompt




The data above consist of a sample of 2,500 peer-to-peer loans issued through the Lending Club (https://www.lendingclub.com/home.action). The interest rate of these loans is determined by the Lending Club on the basis of characteristics of the person asking for the loan such as their employment history, credit history, and creditworthiness scores. 




The purpose of your analysis is to identify and quantify associations between the interest rate of the loan and the other variables in the data set. In particular, you should consider whether any of these variables have an important association with interest rate after taking into account the applicant's FICO score. For example, if two people have the same FICO score, can the other variables explain a difference in interest rate between them?